requirements.txt
pandas==1.2.3
numpy==1.20.1 
pip==18.1   
# pkg-resources==0.0.0   
python-dateutil==2.8.1  
pytz==2021.1 
setuptools==40.8.0 
six==1.15.0 
pytest==6.2.2
psycopg2-binary==2.8.6

boto3==1.17.22
botocore==1.20.22
jmespath==0.10.0
python-dateutil==2.8.1
s3transfer==0.3.4
# six==1.15.0
urllib3==1.26.3


python-dotenv